Make get_user_session break properly
This change will intentionally throw a RecordNotFound-error to prevent a
nil-error later on. The intended behaviour is to log the user out and
redirect to login, but this is currently impossible because helpers
can't redirect in all cases.

The authentication helpers should probably be moved to a concern, but I
can't be bothered enough, so it remains fixme.